typed dataframes (#579)
* Arithmetic functions to typed DF and test

* switch to typed


* replace incorrect uses of shorthand xsd: notation


* remove show


* solve some issues from Construct


* Make more tests pass


* fallback typer to strings


* make BGPSpec pass


* make LcaseSpec pass


* make CoalesceSpec pass


* make BlankNodeSpec pass


* make DescribeSpec pass


* make more tests pass


* fix default graph


* fix strspec


* fix more tests


* fix more tests


* fix more tests


* fix aggregations


* make more tests pass


* Make FuncPropertySpec pass


* Make PropertyPathsSpec pass


* Fix condition for evaluateBGP


* make more tests pass


* Make pathframespec pass


* fix StrstartsSpec


* fix Ask


* fix UUID function


* fix FuncAgg


* fix Graph


* fix Construct


* make more tests pass


* return equals to old implementation


* make VALUES work


* make more tests pass


* remove untyped functions


* fix Describe


* fix MultisetSpec


* fix CompilerSpec


* fix errors in aggregations


* unignore propertypath tests


* lint

* more lint

* fix strlenspec

It tested the same record twice


* unignore PropertyPathsSpec


* fix YearSpec


* ignore non-standard numeric test


* rename FuncForms


* fix more tests


* merge


* make tests pass?


* lint

* remove unneeded if

* fix strafter tests


* create jvmopts for sbt


* try to update JVM settings?


* trying sbtopts instead


* checking if this is the only test failing


* checking if this is the only test failing


* move lint to its own workflow so it runs in parallel

+++ modules/engine/src/main/scala/com/gsk/kg/engine/typed/functions/TypedLiterals.scala
@@ -119,7 +119,7 @@ object TypedLiterals {
       RdfType.Decimal(op(col1.value, col2.value))
     ).when(
       isIntNumericLiteral(col1) || isIntNumericLiteral(col2),
-      RdfType.Int(op(col1.value, col2.value))
+      RdfType.Int(op(col1.value, col2.value).cast(IntegerType))
     )
   // scalastyle:on
